Updated module augmentation for autosave package.
